Moved definition of component to PlayerComponent.h
